// src/redux/type.tsx
export enum Actiontype {
    POPULAR = 'POPULAR',
    GENRE = 'GENRE',
    TRENDING = 'TRENDING',
    TOPRATED = 'TOPRATED',
    UPCOMING = 'UPCOMING',
    ADVENTURE = 'ADVENTURE',
    ANIMATION = 'ANIMATION',
    DETAILMOVIE = 'DETAILMOVIE',
    NEWMOVIE = 'NEWMOVIE',
    SIMILAR = 'SIMILAR',
    CAST = 'CAST',
    TRAILER = 'TRAILER',
    LISTGENRE = 'LISTGENRE',
    DETAILCAST = 'DETAILCAST',
    MOVIECAST = 'MOVIECAST',
    GENRECOUNTRY = 'GENRECOUNTRY',
    LANGUAGE = 'LANGUAGE'
}